+++ migrations/000007_add_app_files_columns.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE venv_build DROP COLUMN IF EXISTS files;
+ALTER TABLE venv_build DROP COLUMN IF EXISTS file_upload_mode;